+++ packages/frontapp/src/app/components/BookInfo/Buttons/DonateButtons.tsx
@@ -1,0 +1,17 @@
+import React, { FC } from 'react';
+import Button from '../../Button';
+import { t } from 'i18next';
+
+interface IEdit {
+  onAccept: () => void;
+  onReject: () => void;
+}
+
+export const DonateButtons: FC<IEdit> = ({ onAccept, onReject }) => {
+  return (
+    <>
+      <Button value={t('Buttons.Accept')} onClick={onAccept} />
+      <Button value={t('Buttons.Reject')} transparent onClick={onReject} />
+    </>
+  );
+};
